+++ sdk/eventhubs/azure_messaging_eventhubs/src/producer/batch.rs
@@ -33,7 +33,7 @@ struct EventDataBatchState {
 /// # use azure_identity::TokenCredentialOptions;
 ///
 /// # async fn send_event_batch() -> Result<(), Box<dyn std::error::Error>> {
-/// # let credentials = azure_identity::DefaultAzureCredential::new()?;
+/// # let credentials = azure_identity::DeveloperToolsCredential::new(None)?;
 /// # let producer_client = ProducerClient::builder()
 /// #     .open("fully_qualified_domain_name", "event_hub_name", credentials.clone()).await?;
 /// #
@@ -163,7 +163,7 @@ impl<'a> EventDataBatch<'a> {
     /// # use azure_messaging_eventhubs::models::EventData;
     ///
     /// # async fn send_event_batch() -> Result<(), Box<dyn std::error::Error>> {
-    /// # let my_credential = azure_identity::DefaultAzureCredential::new()?;
+    /// # let my_credential = azure_identity::DeveloperToolsCredential::new(None)?;
     /// # let producer_client = ProducerClient::builder().open("fully_qualified_domain_name", "event_hub_name", my_credential.clone()).await?;
     /// let mut batch = producer_client.create_batch(None).await?;
     ///
@@ -207,7 +207,7 @@ impl<'a> EventDataBatch<'a> {
     /// # use azure_messaging_eventhubs::models::AmqpMessage;
     ///
     /// # async fn send_event_batch() -> Result<(), Box<dyn std::error::Error>> {
-    /// # let my_credential = azure_identity::DefaultAzureCredential::new()?;
+    /// # let my_credential = azure_identity::DeveloperToolsCredential::new(None)?;
     /// # let producer_client = ProducerClient::builder().open("fully_qualified_domain_name", "event_hub_name", my_credential.clone()).await?;
     /// let mut batch = producer_client.create_batch(None).await?;
     ///