+++ sdk/identity/azure_identity/src/process/mod.rs
@@ -1,6 +1,8 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
+#![cfg_attr(target_arch = "wasm32", allow(unused))]
+
 // cspell:ignore workdir
 
 use crate::env::Env;