+++ sdk/identity/azure_identity/tests/default_options_test.rs
@@ -1,7 +1,9 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-use azure_identity::{AzureCliCredentialOptions, AzureDeveloperCliCredentialOptions};
+use azure_identity::{
+    AzureCliCredentialOptions, AzureDeveloperCliCredentialOptions, DeveloperToolsCredentialOptions,
+};
 
 #[test]
 fn az_credential_options() {
@@ -13,6 +15,13 @@ fn az_credential_options() {
 #[test]
 fn azd_credential_options() {
     let _options = AzureDeveloperCliCredentialOptions {
+        ..Default::default()
+    };
+}
+
+#[test]
+fn developer_tool_credential_options() {
+    let _options = DeveloperToolsCredentialOptions {
         ..Default::default()
     };
 }