+++ sdk/identity/azure_identity/CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ### Breaking Changes
 
+- Replaced `DefaultAzureCredential` with `DeveloperToolsCredential`
+
 ### Bugs Fixed
 
 ### Other Changes